% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauStarDE.R
\name{isProbVector}
\alias{isProbVector}
\title{Check if Vector of Probabilities}
\usage{
isProbVector(probs)
}
\arguments{
\item{probs}{the probability vector to check}
}
\value{
TRUE if conditions are met, FALSE if otherwise
}
\description{
Checks if the input vector has entries that sum to 1 and are non-negative
}
