% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LPS_ori_tbl}
\alias{LPS_ori_tbl}
\title{A tibble containing the inferred pseudotime of \code{LPS_sce} by \code{slingshot} package}
\format{
A tibble with two cols
\describe{
\item{cell}{Cell names}
\item{pseudotime}{Inferred pseudotime of each cell}
}
}
\usage{
data("LPS_ori_tbl")
}
\description{
A tibble containing the inferred pseudotime of \code{LPS_sce} by \code{slingshot} package
}
\keyword{datasets}
