% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LPS_sce}
\alias{LPS_sce}
\title{A SingelCellExperiment dataset containing 4,016 genes and 390 cells}
\format{
A dataset with 4,016 rows and 390 cols
}
\usage{
data("LPS_sce")
}
\description{
A SingelCellExperiment dataset containing 4,016 genes and 390 cells
}
\keyword{datasets}
