% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LPS_sub_tbl}
\alias{LPS_sub_tbl}
\title{A list of tibbles containing the inferred pseudotime of each subsample}
\format{
A list with 1,000 tibbles
}
\usage{
data("LPS_sub_tbl")
}
\description{
A list of tibbles containing the inferred pseudotime of each subsample
}
\keyword{datasets}
