% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUncertainty.R
\name{plotUncertainty}
\alias{plotUncertainty}
\title{Plot the Pseudotime of Subsamples}
\usage{
plotUncertainty(ori.tbl, sub.tbl)
}
\arguments{
\item{ori.tbl}{A tibble or dataframe which contains the original cells and pseudotime as two columns.}

\item{sub.tbl}{A list of tibbles or dataframes where each is the fit of a subsample. Each element is the same format as \code{ori.tbl}.}
}
\value{
A ggplot object
}
\description{
Plot the fitted pseudotime of each cell from all subsamples.
}
\author{
Dongyuan Song, Huy Nyugen
}
