% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotimeDE.R
\name{pseudotimeDE}
\alias{pseudotimeDE}
\title{Perform Differential Expression Test on One Gene}
\usage{
pseudotimeDE(
  gene,
  ori.tbl,
  boot.tbl,
  sce,
  model = c("nb", "zinb", "auto"),
  k = 6,
  knots = c(0:5/5),
  fix.weight = TRUE,
  aicdiff = 10,
  seed = 123
)
}
\arguments{
\item{gene}{A string of gene name. It should be one of the row names in sce.}

\item{ori.tbl}{A tibble or dataframe which contains the original cells and pseudotime as two columns.}

\item{boot.tbl}{A list of tibbles or dataframes where each is the fit of a subsample. Each element is the same format as ori.tbl.}

\item{sce}{A SingleCellExperment object which contain the count data. Its row names should be genes and col names should be cell.}

\item{model}{A string of the model name. One of \code{nb}, \code{zinb} and \code{auto}.}

\item{k}{A integer of the basis dimension. Default is 6.}

\item{knots}{A numeric vector of the position of knots.}

\item{fix.weight}{A logic variable indicating if the ZINB-GAM will use the zero weights from the original model.}

\item{aicdiff}{A numeric variable of the threshold of modle selection. Only works when \code{model = `auto`}.}

\item{seed}{A numeric variable of the random seed. It mainly affects the fitting of null distribution.}
}
\value{
A list with the components:
\describe{
  \item{\code{fix.pv}}{The p-value assuming the pseudotime is fixed}
  \item{\code{permut.pv}}{The permutation p-value}
  \item{\code{smooth.pv}}{The permutation p-value by fitting a parametric null distribution}
  \item{\code{rank}}{The estimated effect degree of freedom of the original model}
  \item{\code{Tr}}{The test statistic of the original model}
  \item{\code{Tr.dist}}{The null sample of the test statistic}
  \item{\code{zinf}}{Whether the model is zero inflated}
  \item{\code{aic}}{The AIC of the orignial model}
  \item{\code{expv.quantile}}{Quantiles of the log counts plus 1}
  \item{\code{expv.mean}}{Mean of the log counts plus 1}
  \item{\code{expv.zero}}{Zero proportions of counts}
}
}
\description{
Test if one gene is differentially expressed with pseudotime.
}
\author{
Dongyuan Song
}
