% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPseudotimeDE.R
\name{runPseudotimeDE}
\alias{runPseudotimeDE}
\title{Perform Differential Expression Test on A List of Genes}
\usage{
runPseudotimeDE(
  gene.vec,
  ori.tbl,
  sub.tbl,
  sce,
  model = c("nb", "zinb", "auto"),
  k = 6,
  knots = c(0:5/5),
  fix.weight = TRUE,
  aicdiff = 10,
  seed = 123,
  mc.cores = 20
)
}
\arguments{
\item{gene.vec}{A vector of genes. It should be a subset of the row names in sce.}

\item{ori.tbl}{A tibble or dataframe which contains the original cells and pseudotime as two columns.}

\item{sub.tbl}{A list of tibbles or dataframes where each is the fit of a subsample. Each element is the same format as ori.tbl.}

\item{sce}{A SingleCellExperment object which contain the count data. Its row names should be genes and col names should be cells.}

\item{model}{A string of the model name. One of \code{nb}, \code{zinb} and \code{auto}.}

\item{k}{A integer of the basis dimension. Default is 6. The reults are usually robust to different k; we recommend to use k from 5 to 10.}

\item{knots}{A numeric vector of the location of knots.}

\item{fix.weight}{A logic variable indicating if the ZINB-GAM will use the zero weights from the original model.}

\item{aicdiff}{A numeric variable of the threshold of modle selection. Only works when \code{model = `auto`}.}

\item{seed}{A numeric variable of the random seed. It mainly affects the fitting of null distribution.}
}
\value{
A tibble of summary results of genes
}
\description{
Test if each gene in a list of genes is differentially expressed along pseudotime.
A wrapper of \code{PseudotimeDE::pseudotimeDE}.
}
\author{
Dongyuan Song
}
