% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauStarDE.R
\name{runTauStarDE}
\alias{runTauStarDE}
\title{Perform Differential Expression Test by tauStar on a List of Genes}
\usage{
runTauStarDE(
  gene.vec,
  ori.tbl,
  mat,
  assay.use = "counts",
  seurat.assay = "RNA",
  mc.cores = 2,
  mc.preschedule = TRUE,
  SIMPLIFY = TRUE
)
}
\arguments{
\item{gene.vec}{A vector of genes. It should be a subset of the row names in sce.}

\item{ori.tbl}{A tibble or dataframe which contains the original cells and pseudotime as two columns.}

\item{mat}{The input expression data. It can be:
(1) A SingleCellExperiment object which contain the expression data;
(2) An matrix;
(3) A Seurat object which contain the expression data.
Its row names should be genes and col names should be cells.}

\item{assay.use}{The \code{assay} used in SingleCellExperiment or \code{slot} used in Seurat. Default is \code{counts}.}

\item{seurat.assay}{The \code{assay} used in Seurat. Default is \code{'RNA'}.}

\item{mc.cores}{Number of cores for computing.}

\item{mc.preschedule}{See \code{mclapply}. Default is TRUE.}

\item{SIMPLIFY}{A logic variable whether to return a tibble (TRUE) or a list of lists (FALSE). Default is TRUE.}
}
\value{
A tibble of summary results of genes
}
\description{
Test if each gene in a list of genes is differentially expressed along pseudotime.
A wrapper of \code{tauStarDE::tauStarDE}.
}
\examples{
data("LPS_sce")
data("LPS_ori_tbl")
data("LPS_sub_tbl")
res <- PseudotimeDE::runTauStarDE(gene.vec = c("CCL5", "CXCL10"),
ori.tbl = LPS_ori_tbl, sub.tbl = LPS_sub_tbl[1:10], mat = LPS_sce, model = "nb")

}
\author{
Yuheng Lai
}
