% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCurve.R
\name{plotCurve}
\alias{plotCurve}
\title{Plot the Fitted Model (Gene Trajectory)}
\usage{
plotCurve(gene.vec, ori.tbl, sce, model.fit, alpha = 0.2, ncol = 2)
}
\arguments{
\item{gene.vec}{A vector of genes. It should be a subset of the row names in sce.}

\item{ori.tbl}{A tibble or dataframe which contains the original cells and pseudotime as two columns.}

\item{sce}{A SingleCellExperment object which contain the count data. Its row names should be genes and col names should be cells.}

\item{model.fit}{A list of fitted gam models corresponding to genes in \code{gene.vec}.}

\item{alpha}{A numeric value of the opacity of points. Default is 0.2.}

\item{ncol}{A integer of facet column number. Default is 2.}
}
\value{
A ggplot object
}
\description{
Plot the fitted curve (gene trajectory) by a given gam model.
}
\author{
Dongyuan Song
}
