% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudotimeDE.R
\name{pseudotimeDE}
\alias{pseudotimeDE}
\title{Perform Differential Expression Test on One Gene}
\usage{
pseudotimeDE(
  gene,
  ori.tbl,
  sub.tbl,
  mat,
  assay.use = "counts",
  model = c("nb", "zinb", "auto"),
  k = 6,
  knots = c(0:5/5),
  fix.weight = TRUE,
  aicdiff = 10,
  seed = 123
)
}
\arguments{
\item{gene}{A string of gene name. It should be one of the row names in sce.}

\item{ori.tbl}{A tibble or dataframe which contains the original cells and pseudotime as two columns.}

\item{sub.tbl}{A list of tibbles or dataframes where each is the fit of a subsample. Each element is the same format as \code{ori.tbl}.}

\item{mat}{The input expression data. It can be:
(1) A SingleCellExperment object which contain the expression data;
(2) An matrix;
(3) A Seurat object which contain the expression data.
Its row names should be genes and col names should be cells.}

\item{assay.use}{The \code{assay} used in SingleCellExperiment or \code{slot} used in Seurat. Default is \code{counts}.}

\item{model}{A string of the model name. One of \code{nb}, \code{zinb} and \code{auto}.}

\item{k}{A integer of the basis dimension. Default is 6. The reults are usually robust to different k; we recommend to use k from 5 to 10.}

\item{knots}{A numeric vector of the location of knots. Default is evenly distributed between 0 to 1.}

\item{fix.weight}{A logic variable indicating if the ZINB-GAM will use the zero weights from the original model. Used for saving time since ZINB-GAM is computationally intense.}

\item{aicdiff}{A numeric variable of the threshold of model selection. Only works when \code{model = `auto`}.}

\item{seed}{A numeric variable of the random seed. It mainly affects the parametricfitting of null distribution.}
}
\value{
A list with the components:
\describe{
  \item{\code{fix.pv}}{The p-value assuming the pseudotime is fixed}
  \item{\code{emp.pv}}{The permutation p-value}
  \item{\code{para.pv}}{The permutation p-value by fitting a parametric null distribution}
  \item{\code{ad.pv}}{P-value of Anderson-Darling test on comparing null distribution and its parametric fit}
  \item{\code{rank}}{The estimated effect degree of freedom of the original model}
  \item{\code{gam.fit}}{The fitted gam model on original data}
  \item{\code{zinf}}{Whether the model is zero inflated}
  \item{\code{aic}}{The AIC of the orignial model}
  \item{\code{expv.quantile}}{Quantiles of the log counts plus 1}
  \item{\code{expv.mean}}{Mean of the log counts plus 1}
  \item{\code{expv.zero}}{Zero proportions of counts}
}
}
\description{
Test if one gene is differentially expressed along pseudotime.
}
\author{
Dongyuan Song
}
