% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPseudotimeDE.R
\name{runPseudotimeDE}
\alias{runPseudotimeDE}
\title{Perform Differential Expression Test on a List of Genes}
\usage{
runPseudotimeDE(
  gene.vec,
  ori.tbl,
  sub.tbl,
  mat,
  assay.use = "counts",
  model = c("nb", "zinb", "gaussian", "auto", "qgam"),
  k = 6,
  knots = c(0:5/5),
  fix.weight = TRUE,
  aicdiff = 10,
  seed = 123,
  quant = 0.5,
  usebam = FALSE,
  formula = NULL,
  seurat.assay = "RNA",
  mc.cores = 2,
  mc.preschedule = TRUE,
  SIMPLIFY = TRUE
)
}
\arguments{
\item{gene.vec}{A vector of genes. It should be a subset of the row names in sce.}

\item{ori.tbl}{A tibble or dataframe which contains the original cells and pseudotime as two columns.}

\item{sub.tbl}{A list of tibbles or dataframes where each is the fit of a subsample. Each element is the same format as ori.tbl.}

\item{mat}{The input expression data. It can be:
(1) A SingleCellExperiment object which contain the expression data;
(2) An matrix;
(3) A Seurat object which contain the expression data.
Its row names should be genes and col names should be cells.}

\item{assay.use}{The \code{assay} used in SingleCellExperiment or \code{slot} used in Seurat. Default is \code{counts}.}

\item{model}{A string of the model name. One of \code{nb}, \code{zinb}, \code{gaussian}, \code{auto} and \code{qgam}.}

\item{k}{A integer of the basis dimension. Default is 6. The results are usually robust to different k; we recommend to use k from 5 to 10.}

\item{knots}{A numeric vector of the location of knots. Default is evenly distributed between 0 to 1. For instance, if your k = 6, and your range is [0, 10], then the position of knots should be \code{c(0:5)*(10-0)}.}

\item{fix.weight}{A logic variable indicating if the ZINB-GAM will use the zero weights from the original model.}

\item{aicdiff}{A numeric variable of the threshold of model selection. Only works when \code{model = `auto`}.}

\item{seed}{A numeric variable of the random seed. It mainly affects the fitting of null distribution.}

\item{quant}{The quantile of interest for quantile regression (qgam), range from 0 to 1, default as 0.5 (median).}

\item{usebam}{A logical variable. If use \code{mgcv::bam}, which may be faster with large sample size (e.g., > 10'000 cells).}

\item{formula}{An (optional) custom formula to be passed to \code{mgcv::gam}, \code{mgcv::bam}, or \code{qgam::qgam}.}

\item{seurat.assay}{The \code{assay} used in Seurat. Default is \code{'RNA'}.}

\item{mc.cores}{Number of cores for computing.}

\item{mc.preschedule}{See \code{mclapply}. Default is TRUE.}

\item{SIMPLIFY}{A logic variable whether to return a tibble (TRUE) or a list of lists (FALSE). Default is TRUE.}
}
\value{
A tibble of summary results of genes
}
\description{
Test if each gene in a list of genes is differentially expressed along pseudotime.
A wrapper of \code{PseudotimeDE::pseudotimeDE}.
}
\examples{
data("LPS_sce")
data("LPS_ori_tbl")
data("LPS_sub_tbl")
res <- PseudotimeDE::runPseudotimeDE(gene.vec = c("CCL5", "CXCL10"),
ori.tbl = LPS_ori_tbl, sub.tbl = LPS_sub_tbl[1:10], mat = LPS_sce, model = "nb")

}
\author{
Dongyuan Song, Shiyu Ma
}
